Improved test coverage + bug fixes (#25)
* Implemented getObject, fixed halt command, fixed fireAllRule command, general cleanup, improved test coverage

* Added timout on tests

+++ kie-remote/src/main/java/org/kie/remote/RemoteStatefulSession.java
@@ -19,7 +19,7 @@ package org.kie.remote;
 import java.util.concurrent.CompletableFuture;
 
 public interface RemoteStatefulSession {
-    CompletableFuture<Integer> fireAllRules();
+    CompletableFuture<Long> fireAllRules();
     void fireUntilHalt();
     void halt();
 }
